pyproject.toml
[tool.poetry]
name = "task_tracker"
version = "0.1.0"
description = ""
authors = ["Rishat Sharafiev <rishatsharafiev@ya.ru>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.7.4"
django = "^3.0.8"
python-decouple = "^3.3"
dj-database-url = "^0.5.0"
celery = "^4.4.6"
django-celery-beat = "^2.0.0"
sentry-sdk = "^0.15.1"
structlog = "^20.1.0"
structlog-sentry = "^1.2.2"
django-structlog = "^1.6.0"
colorama = "^0.4.3"
django-cors-headers = "^3.4.0"
django-debug-toolbar = "^2.2"
flower = "^0.9.4"
Pillow = "^7.2.0"
redis = "^3.5.3"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
